packages/ai_extraction/requirements.txt
openai
PyPDF2
python-docx
python-dotenv
striprtf 